requirements.txt
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
